backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
httpx==0.28.1
anthropic==0.49.0
pinecone==6.0.2
voyageai==0.3.2
pymupdf==1.25.3
python-docx==1.1.2
python-pptx==1.0.2
langchain-text-splitters==0.3.4
sse-starlette==2.2.1
google-auth==2.38.0
